refactor: add ConcentratedLiquidityDepthChart as reusable component
