+++ packages/web/.env
@@ -5,8 +5,8 @@
 # Osmosis Chain Configuration Overwrite
 # NEXT_PUBLIC_IS_FRONTIER=false
 # NEXT_PUBLIC_IS_TESTNET=true
- NEXT_PUBLIC_OSMOSIS_RPC_OVERWRITE=http://localhost:26657/
- NEXT_PUBLIC_OSMOSIS_REST_OVERWRITE=http://localhost:1317/
+# NEXT_PUBLIC_OSMOSIS_RPC_OVERWRITE=http://localhost:26657/
+# NEXT_PUBLIC_OSMOSIS_REST_OVERWRITE=http://localhost:1317/
 # NEXT_PUBLIC_OSMOSIS_RPC_OVERWRITE=https://rpc-v13-devnet.osmosis.zone/
 # NEXT_PUBLIC_OSMOSIS_REST_OVERWRITE=https://lcd-v13-devnet.osmosis.zone/
 # NEXT_PUBLIC_OSMOSIS_CHAIN_ID_OVERWRITE=v13.0.0-rc3-testnet
